Big update.
Added a real type, which is a float but can be made a double with a build
flag. All types/functions should use this going forward. Main purpose is to
easily go to double precision if we need to.
Also added a sprinkle of C++ operator overloading, for my fingers sake.